refactor(openomni): split skill index into focused modules
Extract skill/index.ts (467 LOC) into 5 focused modules:
- shared.ts (67 LOC): path resolution, sort utilities, constants
- registry.ts (48 LOC): SkillRegistry namespace (read/write)
- markdown.ts (162 LOC): SKILL.md parsing and metadata extraction
- descriptors.ts (67 LOC): runtime resource descriptor attachment
- loader.ts (111 LOC): SkillLoader namespace (discover/load)

index.ts becomes a 16 LOC barrel re-export.
